Refactor `formatPath()` so core call sites don’t need to pass config
