+++ packages/starlight/components/SiteTitle.astro
@@ -3,12 +3,8 @@ import { logos } from 'virtual:starlight/user-images';
 import config from 'virtual:starlight/user-config';
 import type { Props } from '../props';
 import { formatPath } from '../utils/format-path';
-import project from 'virtual:starlight/project-context';
 
-const href = formatPath(Astro.props.locale || '/', {
-	format: project.build.format,
-	trailingSlash: project.trailingSlash,
-});
+const href = formatPath(Astro.props.locale || '/');
 ---
 
 <a {href} class="site-title sl-flex">
